Move ScreenAddUser logic to PageAddUser
