+++ src/app/core/navigation/navigation.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import { ScreenAddUser } from '../../modules/add-user/add-user';
+import { PageAddUser } from '../../../atomic/pag.add-user/add-user.component';
 import { ScreenLogin } from '../../modules/login/login';
 import { ScreenUserList } from '../../modules/user-list/user-list';
 
@@ -14,7 +14,7 @@ export const Navigation = () => (
     <Stack.Navigator screenOptions={{ title: '' }} initialRouteName="login">
       <Stack.Screen name="user-list" component={ScreenUserList} />
       <Stack.Screen name="login" component={ScreenLogin} />
-      <Stack.Screen name="add-user" component={ScreenAddUser} />
+      <Stack.Screen name="add-user" component={PageAddUser} />
     </Stack.Navigator>
   </NavigationContainer>
 );
